Basic buy without validation working. Some Refactoring

+++ vending_machine/models.py
@@ -9,7 +9,7 @@ class User(db.Model, UserMixin):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(), unique=True)
     password = db.Column(db.String())
-    deposit = db.Column(db.Float, default=0.0)
+    deposit = db.Column(db.Integer, default=0)
     role = db.Column(db.String(6), default='buyer')
     products_sold = db.relationship('Product', backref='seller')
